token 缓存

+++ api-server/src/main/java/com/wfj/learn/apiserver/entity/User.java
@@ -6,14 +6,14 @@ import lombok.Data;
 @Data
 public class User {
 
-    private Long id;
+    private Integer id;
     private String username;
     private String password;
 
     public User() {
     }
 
-    public User(Long id, String username, String password) {
+    public User(Integer id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
